+++ test/edax_command_test.dart
@@ -7,6 +7,17 @@ import 'package:test/test.dart';
 const _testBookFile = './resources/test_book.dat';
 
 void main() {
+  test('open and close DynamicLibrary', () {
+    final libedax = LibEdax(); // open internally
+    libedax.edaxVersion(); // ignore: cascade_invocations
+
+    libedax.closeDll(); // ignore: cascade_invocations
+    // libedax.edaxVersion(); => CRASH
+
+    final libedax2 = LibEdax(); // open internally
+    libedax2.edaxVersion(); // ignore: cascade_invocations
+  });
+
   group('with a new book (follow default: data/book.dat)', () {
     test('initialize without args, and set option', () {
       LibEdax()